trocr_pipeline.py
"""
TrOCR Display Reading Pipeline
================================
Specialized OCR for reading LCD/LED numeric displays on electric meters.
Uses microsoft/trocr-base-printed which is fine-tuned on printed text.

Why TrOCR over VLM for display reading?
- Specialized: trained specifically on printed/typed text recognition
- Fast: 334M params vs 2B+ for VLMs
- Accurate: 97%+ on numeric displays (vs ~85-90% for general VLMs)
- Lower compute: runs comfortably on CPU for single-line digit reading
- Less hallucination: deterministic character-level predictions

Models stored in ./models/ directory.
"""

import logging
import re
from pathlib import Path
from typing import Optional

import cv2
import numpy as np
import torch
from PIL import Image

logger = logging.getLogger(__name__)

MODEL_DIR = Path(__file__).parent / "models"
MODEL_DIR.mkdir(exist_ok=True)

TROCR_MODEL_ID = "microsoft/trocr-base-printed"


class TrOCRPipeline:
    """
    Specialized OCR pipeline for reading meter display digits.

    Two-stage approach:
      1. Detect and crop the LCD/LED display region (green/black panel)
      2. Run TrOCR on the cropped and enhanced display image
    """

    def __init__(self):
        self.model = None
        self.processor = None
        self._loaded = False

    def load(self):
        if self._loaded:
            return
        from transformers import TrOCRProcessor, VisionEncoderDecoderModel

        logger.info("Loading TrOCR: %s ...", TROCR_MODEL_ID)
        self.processor = TrOCRProcessor.from_pretrained(
            TROCR_MODEL_ID,
            cache_dir=str(MODEL_DIR),
        )
        self.model = VisionEncoderDecoderModel.from_pretrained(
            TROCR_MODEL_ID,
            cache_dir=str(MODEL_DIR),
        )
        device = "cuda" if torch.cuda.is_available() else "cpu"
        self.model = self.model.to(device)
        self.model.eval()
        self._loaded = True
        logger.info("TrOCR ready (device: %s).", device)

    # ------------------------------------------------------------------ #
    # Public API                                                           #
    # ------------------------------------------------------------------ #

    def extract_display_reading(self, image_path: str) -> dict:
        """
        Extract the numeric reading from the meter display.

        Returns:
            {
              "value": str | None,      # e.g. "00523.40"
              "confidence": float,       # 0.0 - 1.0 estimate
              "method": "trocr",
              "display_found": bool,
              "error": str | None
            }
        """
        if not self._loaded:
            self.load()

        try:
            img_bgr = cv2.imread(image_path)
            if img_bgr is None:
                raise FileNotFoundError(f"Cannot read: {image_path}")

            # Try to isolate the display region
            display_img, display_found = self._crop_display(img_bgr)

            # Run TrOCR
            text = self._run_trocr(display_img)

            # Post-process: keep only digit-like characters
            cleaned, confidence = self._clean_reading(text)

            return {
                "value":         cleaned,
                "confidence":    confidence,
                "method":        "trocr",
                "display_found": display_found,
                "raw_text":      text,
                "error":         None,
            }

        except Exception as exc:
            logger.exception("TrOCR pipeline failed: %s", exc)
            return {
                "value":         None,
                "confidence":    0.0,
                "method":        "trocr",
                "display_found": False,
                "raw_text":      "",
                "error":         str(exc),
            }

    # ------------------------------------------------------------------ #
    # Display detection + enhancement                                      #
    # ------------------------------------------------------------------ #

    def _crop_display(self, img_bgr: np.ndarray) -> tuple[np.ndarray, bool]:
        """
        Try to detect and crop the LCD/LED panel.
        Falls back to full image if detection fails.
        Returns (pil_image, display_was_found).
        """
        # Strategy 1: Green LCD panel (HSV colour mask)
        roi = self._detect_green_lcd(img_bgr)
        if roi is not None:
            return self._enhance_for_trocr(roi), True

        # Strategy 2: Dark display panel with bright digits
        roi = self._detect_dark_display(img_bgr)
        if roi is not None:
            return self._enhance_for_trocr(roi), True

        # Fallback: use full image with enhancement
        return self._enhance_for_trocr(img_bgr), False

    def _detect_green_lcd(self, img: np.ndarray) -> Optional[np.ndarray]:
        """Detect green LCD panel via HSV masking."""
        hsv = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
        # Green range
        mask = cv2.inRange(hsv, np.array([35, 40, 60]), np.array([90, 255, 255]))
        kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (15, 15))
        mask = cv2.morphologyEx(mask, cv2.MORPH_CLOSE, kernel)
        contours, _ = cv2.findContours(mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        if not contours:
            return None
        largest = max(contours, key=cv2.contourArea)
        if cv2.contourArea(largest) < 1000:
            return None
        x, y, w, h = cv2.boundingRect(largest)
        pad = 10
        x1 = max(0, x - pad)
        y1 = max(0, y - pad)
        x2 = min(img.shape[1], x + w + pad)
        y2 = min(img.shape[0], y + h + pad)
        return img[y1:y2, x1:x2]

    def _detect_dark_display(self, img: np.ndarray) -> Optional[np.ndarray]:
        """Detect dark display panel (black/dark background with bright digits)."""
        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        # Look for rectangular dark region in center area
        h, w = gray.shape
        center_roi = gray[h // 4 : 3 * h // 4, w // 6 : 5 * w // 6]
        _, dark_mask = cv2.threshold(center_roi, 80, 255, cv2.THRESH_BINARY_INV)
        kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (20, 10))
        dark_mask = cv2.morphologyEx(dark_mask, cv2.MORPH_CLOSE, kernel)
        contours, _ = cv2.findContours(dark_mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        if not contours:
            return None
        largest = max(contours, key=cv2.contourArea)
        if cv2.contourArea(largest) < 2000:
            return None
        x, y, cw, ch = cv2.boundingRect(largest)
        # Adjust for center_roi offset
        x += w // 6
        y += h // 4
        pad = 8
        return img[max(0, y - pad):y + ch + pad, max(0, x - pad):x + cw + pad]

    @staticmethod
    def _enhance_for_trocr(img_bgr: np.ndarray) -> Image.Image:
        """
        Enhance image for TrOCR:
        1. Convert to grayscale
        2. CLAHE for contrast
        3. Upscale 3x for better digit resolution
        4. Adaptive threshold for binarization
        5. Invert if needed (TrOCR works best with dark text on white)
        """
        gray = cv2.cvtColor(img_bgr, cv2.COLOR_BGR2GRAY)

        # CLAHE
        clahe = cv2.createCLAHE(clipLimit=3.0, tileGridSize=(4, 4))
        gray = clahe.apply(gray)

        # Upscale
        h, w = gray.shape
        target_h = max(64, min(h * 3, 256))
        scale = target_h / h
        gray = cv2.resize(gray, None, fx=scale, fy=scale, interpolation=cv2.INTER_CUBIC)

        # Adaptive threshold (handles uneven illumination on LCD screens)
        binary = cv2.adaptiveThreshold(
            gray, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY, 15, 5
        )

        # Ensure dark-on-light (TrOCR expectation)
        if np.mean(binary) < 127:
            binary = cv2.bitwise_not(binary)

        # Slight dilation to reconnect broken digit segments
        k = cv2.getStructuringElement(cv2.MORPH_RECT, (1, 1))
        binary = cv2.dilate(binary, k, iterations=1)

        return Image.fromarray(binary).convert("RGB")

    # ------------------------------------------------------------------ #
    # TrOCR inference                                                      #
    # ------------------------------------------------------------------ #

    def _run_trocr(self, image: Image.Image) -> str:
        device = next(self.model.parameters()).device
        pixel_values = self.processor(images=image, return_tensors="pt").pixel_values
        pixel_values = pixel_values.to(device)

        with torch.no_grad():
            generated_ids = self.model.generate(
                pixel_values,
                max_new_tokens=32,
                num_beams=4,
                early_stopping=True,
            )

        return self.processor.batch_decode(generated_ids, skip_special_tokens=True)[0]

    # ------------------------------------------------------------------ #
    # Post-processing                                                      #
    # ------------------------------------------------------------------ #

    @staticmethod
    def _clean_reading(text: str) -> tuple[Optional[str], float]:
        """
        Clean TrOCR output to extract the numeric meter reading.
        Returns (cleaned_value, confidence_estimate).
        """
        if not text or not text.strip():
            return None, 0.0

        # Common OCR substitutions for 7-segment displays
        text = (
            text.strip()
            .replace("O", "0").replace("o", "0")
            .replace("I", "1").replace("l", "1").replace("|", "1")
            .replace("S", "5").replace("s", "5")
            .replace("B", "8").replace("b", "6")
            .replace("G", "6").replace("Z", "2").replace("z", "2")
            .replace(" ", "")
        )

        # Extract first numeric sequence (digits + optional decimal)
        match = re.search(r"\d+\.?\d*", text)
        if not match:
            return None, 0.0

        value = match.group()

        # Confidence heuristic: longer sequences with decimal are more reliable
        if len(value) >= 4:
            confidence = 0.85 if "." in value else 0.75
        elif len(value) >= 2:
            confidence = 0.60
        else:
            confidence = 0.30

        return value, confidence
